Client/requirements.txt
pygame
RPi.GPIO
picamera2
requests
python-dotenv